FE-664: move mock API responses out of __integration__
