+++ __mocks__/axios.js
@@ -1,4 +1,4 @@
-import lookup from 'src/__integration__/http-responses';
+import lookup from 'src/__testHelpers__/http-responses';
 
 // lookup() returns an API response body object or throws on error to simulate axios exception behaviour
 export const singleton = jest.fn((request) => new Promise((resolve) => resolve(lookup(request))));
