+++ src/__testHelpers__/http-responses/default/v1/account/cors-data/post.js
