+++ src/__testHelpers__/http-responses/default/v1/authenticate/grants/get.js
