+++ src/__testHelpers__/http-responses/default/v1/authenticate/post.js
