+++ src/__testHelpers__/http-responses/default/v1/events/message/get.js
