+++ src/__testHelpers__/http-responses/default/v1/users/sso-username/saml/get.js
