+++ src/__testHelpers__/http-responses/default/v1/users/test-username/saml/get.js
